Moved some functionality out of the utils folder, into apps where they make more sense to be
